src/gov/nist/cscrm/pyproject.toml
[tool.poetry]
name = "cscrm"
version = "0.1.0"
description = "A tool to transform SP 800-161 Rev. 1 Appendix controls from the PDF to OSCAL profiles and catalogs."
authors = ["Al S <xee5ch@member.fsf.org>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
tabula-py = "^2.7.0"
tabulate = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
